tool/requirements.txt
pyinstaller==6.11.0
PyQt5==5.15.6
pytest==8.3.3
angr==9.2.102
pyvex
sqlalchemy
pycrypto
